Deployment/scripts/index_scripts/requirements.txt
azure-storage-file-datalake
# langchain
openai
pypdf
# tiktoken
azure-identity
azure-ai-textanalytics
azure-search-documents==11.6.0b3
azure-keyvault-secrets
pandas
pyodbc